# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: beelms API
  version: 0.1.0
  license:
    name: UNLICENSED
    url: https://example.com/unlicense
  description: |
    OpenAPI спецификация за основните MVP услуги на beelms.
servers:
  - url: https://api.beelms.com/api
    description: Production (примерен домейн, с /api префикс)
  - url: http://127.0.0.1:3000/api
    description: Local development (с /api префикс)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminGrantCourseAccessInput:
      type: object
      properties:
        userId:
          type: string
        enroll:
          type: boolean
          nullable: true
        grantReason:
          type: string
          nullable: true
      required: [userId]
...
    CourseSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        language:
          type: string
          example: bg
        status:
          type: string
          description: Статус на курса (draft, active, inactive)
          enum: [draft, active, inactive]
        isPaid:
          type: boolean
          description: Дали курсът е платен (изисква плащане преди enrollment)
        currency:
          type: string
          nullable: true
        priceCents:
          type: integer
          nullable: true
        categoryId:
          type: string
          nullable: true
        category:
          oneOf:
            - $ref: '#/components/schemas/CourseCategorySummary'
            - type: object
              nullable: true
    CourseCategorySummary:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
    CourseCategory:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        order:
          type: integer
        active:
          type: boolean
    CourseModuleItem:
      type: object
      properties:
        id:
          type: string
        itemType:
          type: string
          description: Тип на елемента в програмата на курса
          enum: [wiki, task, quiz]
        title:
          type: string
        order:
          type: integer
        wikiSlug:
          type: string
          nullable: true
        taskId:
          type: string
          nullable: true
        quizId:
          type: string
          nullable: true
    CourseDetail:
      allOf:
        - $ref: '#/components/schemas/CourseSummary'
        - type: object
          properties:
            curriculum:
              type: array
              items:
                $ref: '#/components/schemas/CourseModuleItem'
...
security:
  - bearerAuth: []
paths:
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [Courses]
      summary: Детайл на курс
      security: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детайл на курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "404":
          description: Курсът не е намерен
...
  /admin/course-categories:
    get:
      operationId: listAdminCourseCategories
      tags: [Admin, Courses]
      summary: Списък с категории за курсове (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Списък с категории
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseCategory'
        "401":
          description: Неоторизиран достъп

    post:
      operationId: createAdminCourseCategory
      tags: [Admin, Courses]
      summary: Създаване на категория за курсове
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slug:
                  type: string
                title:
                  type: string
                order:
                  type: integer
                active:
                  type: boolean
              required: [slug, title]
      responses:
        "201":
          description: Категорията е създадена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseCategory'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
  /admin/course-categories/{id}:
    patch:
      operationId: updateAdminCourseCategory
      tags: [Admin, Courses]
      summary: Редакция на категория за курсове
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slug:
                  type: string
                title:
                  type: string
                order:
                  type: integer
                active:
                  type: boolean
      responses:
        "200":
          description: Обновена категория
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseCategory'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
  /admin/courses:
    get:
      operationId: listAdminCourses
      tags: [Admin, Courses]
      summary: Списък с курсове за администрация
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Списък с курсове
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseSummary'
        "401":
          description: Неоторизиран достъп

    post:
      operationId: createAdminCourse
      tags: [Admin, Courses]
      summary: Създаване на нов курс
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                language:
                  type: string
                status:
                  type: string
                  enum: [draft, active, inactive]
                isPaid:
                  type: boolean
                currency:
                  type: string
                  nullable: true
                priceCents:
                  type: integer
                  nullable: true
                categoryId:
                  type: string
                  nullable: true
              required: [title, description, language, status]
      responses:
        "201":
          description: Курсът е създаден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
  /admin/courses/{courseId}:
    get:
      operationId: getAdminCourse
      tags: [Admin, Courses]
      summary: Детайл на курс за администрация
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детайл на курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "401":
          description: Неоторизиран достъп
        "404":
          description: Курсът не е намерен

    patch:
      operationId: updateAdminCourse
      tags: [Admin, Courses]
      summary: Редакция на курс
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                language:
                  type: string
                status:
                  type: string
                  enum: [draft, active, inactive]
                isPaid:
                  type: boolean
                currency:
                  type: string
                  nullable: true
                priceCents:
                  type: integer
                  nullable: true
                categoryId:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Обновен курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
        "404":
          description: Курсът не е намерен
  /admin/courses/{courseId}/grants:
    post:
      operationId: grantAdminCourseAccess
      tags: [Admin, Courses]
      summary: Grant access до paid course (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminGrantCourseAccessInput'
      responses:
        "204":
          description: Access granted
        "400":
          description: Невалидни данни
        "401":
          description: Неоторизиран достъп
        "404":
          description: Курсът или потребителят не са намерени
...